Systeme de frames finalisé. Rectification mineure tir secondaire. Ajout conditions déblocage tir : tir secondaire => score > 100, tir tertiaire => score > 1000.

+++ include/GraphicElement.h
@@ -12,7 +12,7 @@ private:
     int _id;
     int _w, _h;
     int _nbFrames;
-    int _currentFrame;
+    float _currentFrame;
     bool _perpetual;
 
 public:
